requirements.txt
streamlit
streamlit-authenticator==0.2.3
PyYAML